Separate bibtex and latex run things into packages
