+++ src/nl/hannahsten/texifyidea/run/compiler/BibtexCompiler.kt
@@ -2,7 +2,7 @@ package nl.hannahsten.texifyidea.run.compiler
 
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.roots.ProjectRootManager
-import nl.hannahsten.texifyidea.run.BibtexRunConfiguration
+import nl.hannahsten.texifyidea.run.bibtex.BibtexRunConfiguration
 import nl.hannahsten.texifyidea.util.LatexDistribution
 
 /**
