+++ src/nl/hannahsten/texifyidea/run/evince/EvinceForwardSearch.kt
@@ -2,7 +2,7 @@ package nl.hannahsten.texifyidea.run.evince
 
 import com.intellij.execution.process.ProcessHandler
 import com.intellij.execution.runners.ExecutionEnvironment
-import nl.hannahsten.texifyidea.run.LatexRunConfiguration
+import nl.hannahsten.texifyidea.run.latex.LatexRunConfiguration
 import nl.hannahsten.texifyidea.util.caretOffset
 import nl.hannahsten.texifyidea.util.openedEditor
 import nl.hannahsten.texifyidea.util.psiFile
